feat(shltr): adopt the new design for chrome and dashboard
First pass at bringing the client's Lovable mockup into the shltr theme.
Scope is the app chrome (sidebar + new top bar) and the dashboard, which is
the page the mockup details in full. theme-default is untouched.

Add Tailwind v4 alongside Bootstrap, with two deliberate deviations:

- Every utility is prefixed `tw:`. Bootstrap and Tailwind collide across the
  whole spacing scale (.p-3, .mt-4, .gap-3) with *different* values, plus
  .border/.rounded/.text-center. Bootstrap's utilities are emitted unscoped
  into the shared bundle, so an unprefixed collision would have hit
  theme-default as well.
- Utilities are imported without layer(utilities), and preflight is left out.
  Unlayered CSS beats layered CSS regardless of specificity, so inside
  @layer utilities every `tw:` class silently loses to bare Bootstrap element
  rules; Bootstrap's own reset stays the only reset.

Markup diverges between the two designs via the switcher pattern already
established in views/equipment/equipment_view: a thin component picks the
variant and the logic lives in a shared mixin. Note that Vue 3 does not merge
setup() from mixins, so the new mixins expose stores through computed
properties instead.

- theme resolution moves from main.js into src/theme.js, with isShltrTheme
  added to the global componentMixin; the two existing classList checks in
  EquipmentView/LocationView now use it
- sidebar splits into the_nav/NavShltr + NavDefault + navMixin; NavItems*
  keep their permission logic and are restyled from shltr.scss
- new TheTopBar carries the page title and the unaccepted-orders bell; the
  teal gradient band becomes a plain white strip under shltr
- Dashboard splits into dashboard_view/DashboardShltr + DashboardDefault +
  dashboardMixin, with DashboardBlockShltr and BranchPhotoCardShltr card
  primitives; the eight repeated KPI blocks become one loop

DashboardDefault's template is byte-identical to the previous Dashboard
template, and every rule in theme-default.css remains scoped to
.theme-default.

Known gaps, deliberately left for a follow-up: the mockup's top-bar search is
omitted (this app has no global search, only per-view SearchForm), the page
title is duplicated on views that render their own header, and mobile is
unaddressed as before.
