+++ package.json
@@ -47,6 +47,7 @@
   "devDependencies": {
     "@iconify-json/bi": "^1.2.7",
     "@intlify/eslint-plugin-vue-i18n": "^4.5.1",
+    "@tailwindcss/vite": "^4.3.3",
     "@vitejs/plugin-vue": "^6.0.4",
     "@vuelidate/core": "^2.0.3",
     "@vuelidate/validators": "^2.0.4",
@@ -55,6 +56,7 @@
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-vue": "^10.8.0",
     "sass": "^1.49.8",
+    "tailwindcss": "^4.3.3",
     "unplugin-icons": "^22.5.0",
     "unplugin-vue-components": "^30.0.0",
     "vite": "7.3.1",
