+++ src/components/TheAppLayout.vue
@@ -3,6 +3,8 @@
     <TheNavLoggedIn v-if="store.isLoggedIn" />
     <TheNavLoggedOut v-if="!store.isLoggedIn" />
 
+    <TheTopBar v-if="store.isLoggedIn && isShltrTheme" />
+
     <router-view :key="$route.fullPath" name="app-content" v-slot="{ Component }">
       <component :is="Component" v-bind="props" />
     </router-view>
@@ -16,6 +18,8 @@ import {useRoute} from 'vue-router'
 import {useAuthStore} from "@/stores/auth";
 import TheNavLoggedIn from './TheNavLoggedIn.vue'
 import TheNavLoggedOut from './TheNavLoggedOut.vue'
+import TheTopBar from './TheTopBar.vue'
+import {isShltrTheme} from '@/theme'
 
 const store = useAuthStore()
 const route = useRoute()
