+++ src/components/TheAppLayoutNoSubNav.vue
@@ -3,6 +3,8 @@
     <TheNavLoggedIn v-if="isLoggedIn" />
     <TheNavLoggedOut v-if="!isLoggedIn" />
 
+    <TheTopBar v-if="isLoggedIn && isShltrTheme" />
+
     <div class="container main-app-container">
       <router-view :key="$route.fullPath" name="app-content"></router-view>
     </div>
@@ -13,6 +15,7 @@
 
 import TheNavLoggedIn from './TheNavLoggedIn.vue'
 import TheNavLoggedOut from './TheNavLoggedOut.vue'
+import TheTopBar from './TheTopBar.vue'
 import componentMixin from "@/mixins/common";
 
 export default {
@@ -20,6 +23,7 @@ export default {
   components: {
     TheNavLoggedIn,
     TheNavLoggedOut,
+    TheTopBar,
   },
 }
 </script>
